Clean up reference page models
