+++ site/models/reference-extensions.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferenceExtensionsPage extends ReferenceSectionPage
 {
 }
