+++ site/models/reference-helpers.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferenceHelpersPage extends ReferenceSectionPage
 {
 }
