+++ site/models/reference-kirbytags.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferenceKirbytagsPage extends ReferenceSectionPage
 {
 }
