+++ site/models/reference-panelsections.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferencePanelSectionsPage extends ReferenceSectionPage
 {
 }
