+++ site/models/reference-roots.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferenceRootsPage extends ReferenceSectionPage
 {
 }
