+++ site/models/reference-urls.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferenceUrlsPage extends ReferenceSectionPage
 {
 }
