+++ site/models/reference.php
@@ -1,7 +1,5 @@
 <?php
 
-use Kirby\Reference\ReferenceSectionPage;
-
 class ReferencePage extends ReferenceSectionPage
 {
 }
